Engine: Updating the correlation. Adding data extraction prior to correlation.
