+++ PlatformEngine/src/main/java/com/cognizant/devops/platformengine/app/Application.java
@@ -29,7 +29,6 @@ import com.cognizant.devops.platformcommons.config.ApplicationConfigCache;
 import com.cognizant.devops.platformcommons.config.ApplicationConfigProvider;
 import com.cognizant.devops.platformengine.modules.aggregator.EngineAggregatorModule;
 import com.cognizant.devops.platformengine.modules.correlation.EngineCorrelatorModule;
-import com.cognizant.devops.platformengine.modules.dataextractor.EngineDataExtractorModule;
 import com.cognizant.devops.platformengine.modules.mapper.ProjectMapperModule;
 
 /**
@@ -95,19 +94,6 @@ public class Application {
 						.withIntervalInSeconds(defaultInterval)
 						.repeatForever())
 				.build();
-		
-		// Schedule the data extraction Module.
-		JobDetail dataExtractionJob = JobBuilder.newJob(EngineDataExtractorModule.class)
-				.withIdentity("EngineDataExtractorModule", "iSight")
-				.build();
-
-		Trigger dataExtractionTrigger = TriggerBuilder.newTrigger()
-				.withIdentity("EngineDataExtractorModuleTrigger", "iSight")
-				.startNow()
-				.withSchedule(SimpleScheduleBuilder.simpleSchedule()
-						.withIntervalInSeconds(defaultInterval)
-						.repeatForever())
-				.build();
 
 		// Tell quartz to schedule the job using our trigger
 		Scheduler scheduler;
@@ -117,7 +103,6 @@ public class Application {
 			scheduler.scheduleJob(aggrgatorJob, aggregatorTrigger);
 			scheduler.scheduleJob(correlationJob, correlationTrigger);
 			scheduler.scheduleJob(projectMappingJob, projectMappingTrigger);
-			scheduler.scheduleJob(dataExtractionJob, dataExtractionTrigger);
 		} catch (SchedulerException e) {
 			log.error(e);
 		}
